In build_helper, always use INSTALLER, disable fastestmirror for centos/rhel.
Instead of calling apt-get or yum/dnf, always use the INSTALLER
variable so that we get the proper arguments.

For CentOS/RHEL disable the fastestmirror plugin which takes so long
to determine the fastest mirror that it obliterates any speed up that
would be gained from using that faster mirror.